Remove support for ClojureScript
While ClojureScript support is useful, it is much harder to maintain,
and half of the current issues on the project are related to
ClojureScript support.

I don't like removing functionality, even pre-1.0, but I've reached the
conclusion that a hashp for ClojureScript would be better as a separate
project. Those interested in mantaining such a library can fork this
commit's parent.